% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_GDM_prediction.r
\name{generate_GDM_prediction}
\alias{generate_GDM_prediction}
\title{Generates a prediction for a genetically local area
for a selected species and location}
\usage{
generate_GDM_prediction(selected_location, gdm_model_file)
}
\arguments{
\item{selected_location}{-- a selected point [Required]}

\item{gdm_model_file}{-- an R object bundling everything needed for model [Required]}
}
\value{
a raster containing the prediction of genetically local area
}
\description{
Generates a prediction for a genetically local area
for a selected species and location
}
\examples{
\dontrun{
generate_GDM_prediction <- function( c(150.45, -33.66), CallSerr_genetic_model.Rd )
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
