% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_GDM_prediction.r
\name{generate_GDM_prediction}
\alias{generate_GDM_prediction}
\title{Generates a prediction for a genetically local area
for a selected species and location}
\usage{
generate_GDM_prediction(longitude, latitude, gdm_model_file = NULL,
  species_name = NULL, model_dir = NULL, gdm_model = NULL,
  out_dir = NULL)
}
\arguments{
\item{longitude}{-- selected longitude}

\item{latitude}{-- selected latitude}

\item{gdm_model_file}{-- an R object file containing a model list (called "md")}

\item{species_name}{-- name of species (as it appears in a gdm model file, which ends "_genetic_model.Rd"}

\item{model_dir}{-- directory in which gdm model can be found}

\item{gdm_model}{-- an R list bundling everything required to predict under a GDM}

\item{out_dir}{-- location to place raster of results (if NULL, return raster object)}
}
\value{
a raster containing the prediction of genetically local area
}
\description{
Generates a prediction for a genetically local area
for a selected species and location
}
\examples{
\dontrun{
generate_GDM_prediction <- function( longitude=150.45, latitude=-33.66, gdm_model_file="CallSerr_genetic_model.Rd", out_dir )
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
